+++ data/types/upload/test/client.go
@@ -3,15 +3,15 @@ package test
 import (
 	dataTest "github.com/tidepool-org/platform/data/test"
 	dataTypesUpload "github.com/tidepool-org/platform/data/types/upload"
+	netTest "github.com/tidepool-org/platform/net/test"
 	"github.com/tidepool-org/platform/pointer"
 	"github.com/tidepool-org/platform/test"
-	testInternet "github.com/tidepool-org/platform/test/internet"
 )
 
 func NewClient() *dataTypesUpload.Client {
 	datum := dataTypesUpload.NewClient()
-	datum.Name = pointer.FromString(testInternet.NewReverseDomain())
-	datum.Version = pointer.FromString(testInternet.NewSemanticVersion())
+	datum.Name = pointer.FromString(netTest.RandomReverseDomain())
+	datum.Version = pointer.FromString(netTest.RandomSemanticVersion())
 	datum.Private = dataTest.NewBlob()
 	return datum
 }