+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.29.0
 
+* Refactor net test package
 * Add meta to Dexcom data to assist with debugging validation errors
 * Update Ginkgo suite usage
 * Add optional provider session id to data source update